inlines: relax "left-flanking delimiter run" for superscript/subscript.
We permit punctuation immediately following a delimiter (still omitting
whitespace).
